Vendor files moved to vendor folder in lib
